adding authorization"

+++ app/categories/controller.js
@@ -2,6 +2,15 @@ const Category = require('./model');
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('create', 'Category')) {
+      return res.json({
+        error: 1,
+        message: 'Anda tidak memiliki akses untuk membuat kategori',
+      });
+    }
+
     let payload = req.body;
 
     let category = new Category(payload);
@@ -24,6 +33,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('update', 'Category')) {
+      return res.json({
+        error: 1,
+        message: 'Anda tidak memiliki akses untuk membuat kategori',
+      });
+    }
+
     let payload = req.body;
 
     let category = Category.findOneAndUpdate({ _id: req.params.id }, payload, {
@@ -47,6 +65,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('delete', 'Category')) {
+      return res.json({
+        error: 1,
+        message: 'Anda tidak memiliki akses untuk membuat kategori',
+      });
+    }
+
     let deleted = await Category.findOneAndDelete({ _id: req.params.id });
 
     return res.json(deleted);
