+++ app/tags/controller.js
@@ -2,6 +2,15 @@ const Tag = require('./model');
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('create', 'Tag')) {
+      return res.json({
+        error: 1,
+        message: 'Anda tidak memiliki akses untuk membuat kategori',
+      });
+    }
+
     let payload = req.body;
 
     let tag = new Tag(payload);
@@ -23,6 +32,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('update', 'Tag')) {
+      return res.json({
+        error: 1,
+        message: 'Anda tidak memiliki akses untuk membuat kategori',
+      });
+    }
+
     const payload = req.body;
 
     let tag = await Tag.findOneAndUpdate({ _id: req.params.id }, payload, {
@@ -45,6 +63,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can('delete', 'Tag')) {
+      return res.json({
+        error: 1,
+        message: 'Anda tidak memiliki akses untuk membuat kategori',
+      });
+    }
+
     let deletedTag = await Tag.findOneAndDelete({ _id: req.params.id });
 
     return res.json(deletedTag);
